requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
pydantic==2.10.4
pyyaml>=6.0
python-dotenv>=1.0.0

# Authentication (v5.0)
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Stripe Payment (v5.9.2) — keep aligned with requirements-fly.txt
stripe>=8.0.0

# Document parsing (Docling - IBM)
docling>=2.80.0
PyPDF2==3.0.1
python-docx==1.1.2

# v7.5.0 — additional format handlers + image OCR
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
openpyxl>=3.1.0
python-pptx>=0.6.23
beautifulsoup4>=4.12.0
striprtf>=0.0.26

# Vector search (RAGFlow-inspired)
chromadb>=1.5.0

# Google Drive BYOS (v7.0.0)
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.140.0
cryptography>=42.0.0
